milestone 3

+++ reset.js
@@ -3,16 +3,21 @@ let state_info = JSON.parse(fs.readFileSync('data/data.json'))
 
 // delete micropages
 for(state in state_info){
-  fs.unlinkSync(`public/${getBetterFileName(state)}.html`)
+  if(fs.existsSync(`public/${getBetterFileName(state)}.html`)){
+    fs.unlinkSync(`public/${getBetterFileName(state)}.html`)
+  }
 }
 
 copy('about.html')
 copy('nav.js')
 copy('styles.css')
+copy('about_styles.css')
 copy('IMG_5339.jpeg','images/')
 copy('IMG_5340.jpeg','images/')
 copy('IMG_5341.jpeg','images/')
 copy('IMG_5342.jpeg','images/')
+copy('IMG_0234.png','images/')
+copy('IMG_0235.png','images/')
 
 
 function getBetterFileName(stateName){
@@ -24,6 +29,8 @@ function getBetterFileName(stateName){
 }
 
 function copy(fileName, path=''){
-  fs.unlinkSync('public/'+fileName)
+  if(fs.existsSync('public/'+fileName)){
+    fs.unlinkSync('public/'+fileName)
+  }
   fs.writeFileSync('public/'+fileName, fs.readFileSync('src/'+path+fileName))
 }
